add support of external_reply in RawMessage
