+++ tgbotapi.core/src/commonMain/kotlin/dev/inmo/tgbotapi/types/Contact.kt
@@ -16,4 +16,4 @@ data class Contact(
     val userId: UserId? = null,
     @SerialName(vcardField)
     override val vcard: String? = null
-) : CommonContactData, ExternalReplyInfo.ContentVariant
+) : CommonContactData, ReplyInfo.External.ContentVariant
