+++ tgbotapi.core/src/commonMain/kotlin/dev/inmo/tgbotapi/types/files/AnimationFile.kt
@@ -24,4 +24,4 @@ data class AnimationFile(
     @SerialName(fileSizeField)
     override val fileSize: Long? = null
 ) : TelegramMediaFile, MimedMediaFile, ThumbedMediaFile, PlayableMediaFile, CustomNamedMediaFile, SizedMediaFile,
-    ExternalReplyInfo.ContentVariant
+    ReplyInfo.External.ContentVariant
