+++ tgbotapi.core/src/commonMain/kotlin/dev/inmo/tgbotapi/types/files/AudioFile.kt
@@ -28,6 +28,6 @@ data class AudioFile(
     @SerialName(thumbnailField)
     override val thumbnail: PhotoSize? = null
 ) : TelegramMediaFile, CustomNamedMediaFile, MimedMediaFile, ThumbedMediaFile, PlayableMediaFile, TitledMediaFile,
-    Performerable, ExternalReplyInfo.ContentVariant
+    Performerable, ReplyInfo.External.ContentVariant
 
 fun AudioFile.asVoiceFile() = VoiceFile(fileId, fileUniqueId, duration, mimeType, fileSize)
