+++ tgbotapi.core/src/commonMain/kotlin/dev/inmo/tgbotapi/types/files/DocumentFile.kt
@@ -20,7 +20,7 @@ data class DocumentFile(
     override val mimeType: MimeType? = null,
     @SerialName(fileNameField)
     override val fileName: String? = null
-) : TelegramMediaFile, MimedMediaFile, ThumbedMediaFile, CustomNamedMediaFile, ExternalReplyInfo.ContentVariant
+) : TelegramMediaFile, MimedMediaFile, ThumbedMediaFile, CustomNamedMediaFile, ReplyInfo.External.ContentVariant
 
 @Suppress("NOTHING_TO_INLINE")
 inline fun TelegramMediaFile.asDocumentFile() = if (this is DocumentFile) {
