+++ tgbotapi.core/src/commonMain/kotlin/dev/inmo/tgbotapi/types/games/Game.kt
@@ -2,7 +2,7 @@ package dev.inmo.tgbotapi.types.games
 
 import dev.inmo.tgbotapi.abstracts.TextedInput
 import dev.inmo.tgbotapi.abstracts.Titled
-import dev.inmo.tgbotapi.types.ExternalReplyInfo
+import dev.inmo.tgbotapi.types.ReplyInfo
 import dev.inmo.tgbotapi.types.message.textsources.TextSourcesList
 import dev.inmo.tgbotapi.types.files.AnimationFile
 import dev.inmo.tgbotapi.types.files.Photo
@@ -16,4 +16,4 @@ data class Game(
     override val text: String? = null,
     override val textSources: TextSourcesList = emptyList(),
     val animation: AnimationFile? = null
-) : Titled, TextedInput, ExternalReplyInfo.ContentVariant
+) : Titled, TextedInput, ReplyInfo.External.ContentVariant
