+++ tgbotapi.core/src/commonMain/kotlin/dev/inmo/tgbotapi/types/giveaway/ScheduledGiveaway.kt
@@ -30,4 +30,4 @@ data class ScheduledGiveaway(
     val countries: List<IetfLang>? = null,
     @SerialName(premiumSubscriptionMonthCountField)
     override val premiumMonths: Int? = null
-) : GiveawayInfo, ExternalReplyInfo.ContentVariant
+) : GiveawayInfo, ReplyInfo.External.ContentVariant
