+++ tgbotapi.core/src/commonMain/kotlin/dev/inmo/tgbotapi/types/message/ChannelContentMessageImpl.kt
@@ -16,7 +16,7 @@ data class ChannelContentMessageImpl<T: MessageContent>(
     override val editDate: DateTime?,
     override val hasProtectedContent: Boolean,
     override val forwardOrigin: MessageOrigin?,
-    override val replyTo: AccessibleMessage?,
+    override val replyInfo: ReplyInfo?,
     override val replyMarkup: InlineKeyboardMarkup?,
     override val senderBot: CommonBot?,
     override val authorSignature: AuthorSignature?,
