+++ tgbotapi.core/src/commonMain/kotlin/dev/inmo/tgbotapi/types/message/GroupMessages.kt
@@ -17,7 +17,7 @@ data class ConnectedFromChannelGroupContentMessageImpl<T : MessageContent>(
     override val forwardOrigin: MessageOrigin?,
     override val editDate: DateTime?,
     override val hasProtectedContent: Boolean,
-    override val replyTo: AccessibleMessage?,
+    override val replyInfo: ReplyInfo?,
     override val replyMarkup: InlineKeyboardMarkup?,
     override val content: T,
     override val senderBot: CommonBot?,
@@ -52,7 +52,7 @@ data class UnconnectedFromChannelGroupContentMessageImpl<T: MessageContent>(
     override val forwardOrigin: MessageOrigin?,
     override val editDate: DateTime?,
     override val hasProtectedContent: Boolean,
-    override val replyTo: AccessibleMessage?,
+    override val replyInfo: ReplyInfo?,
     override val replyMarkup: InlineKeyboardMarkup?,
     override val content: T,
     override val senderBot: CommonBot?,
@@ -85,7 +85,7 @@ data class AnonymousGroupContentMessageImpl<T : MessageContent>(
     override val forwardOrigin: MessageOrigin?,
     override val editDate: DateTime?,
     override val hasProtectedContent: Boolean,
-    override val replyTo: AccessibleMessage?,
+    override val replyInfo: ReplyInfo?,
     override val replyMarkup: InlineKeyboardMarkup?,
     override val content: T,
     override val senderBot: CommonBot?,
@@ -118,7 +118,7 @@ data class CommonGroupContentMessageImpl<T : MessageContent>(
     override val forwardOrigin: MessageOrigin?,
     override val editDate: DateTime?,
     override val hasProtectedContent: Boolean,
-    override val replyTo: AccessibleMessage?,
+    override val replyInfo: ReplyInfo?,
     override val replyMarkup: InlineKeyboardMarkup?,
     override val content: T,
     override val senderBot: CommonBot?,
@@ -151,7 +151,7 @@ data class FromChannelForumContentMessageImpl<T: MessageContent>(
     override val forwardOrigin: MessageOrigin?,
     override val editDate: DateTime?,
     override val hasProtectedContent: Boolean,
-    override val replyTo: AccessibleMessage?,
+    override val replyInfo: ReplyInfo?,
     override val replyMarkup: InlineKeyboardMarkup?,
     override val content: T,
     override val senderBot: CommonBot?,
@@ -186,7 +186,7 @@ data class AnonymousForumContentMessageImpl<T : MessageContent>(
     override val forwardOrigin: MessageOrigin?,
     override val editDate: DateTime?,
     override val hasProtectedContent: Boolean,
-    override val replyTo: AccessibleMessage?,
+    override val replyInfo: ReplyInfo?,
     override val replyMarkup: InlineKeyboardMarkup?,
     override val content: T,
     override val senderBot: CommonBot?,
@@ -221,7 +221,7 @@ data class CommonForumContentMessageImpl<T : MessageContent>(
     override val forwardOrigin: MessageOrigin?,
     override val editDate: DateTime?,
     override val hasProtectedContent: Boolean,
-    override val replyTo: AccessibleMessage?,
+    override val replyInfo: ReplyInfo?,
     override val replyMarkup: InlineKeyboardMarkup?,
     override val content: T,
     override val senderBot: CommonBot?,
