+++ tgbotapi.core/src/commonMain/kotlin/dev/inmo/tgbotapi/types/message/abstracts/PossiblyReplyMessage.kt
@@ -1,5 +1,9 @@
 package dev.inmo.tgbotapi.types.message.abstracts
 
+import dev.inmo.tgbotapi.types.ReplyInfo
+
 interface PossiblyReplyMessage {
+    val replyInfo: ReplyInfo?
     val replyTo: AccessibleMessage?
+        get() = (replyInfo as? ReplyInfo.Internal) ?.message
 }
