+++ tgbotapi.core/src/commonMain/kotlin/dev/inmo/tgbotapi/types/stories/Story.kt
@@ -1,7 +1,7 @@
 package dev.inmo.tgbotapi.types.stories
 
-import dev.inmo.tgbotapi.types.ExternalReplyInfo
+import dev.inmo.tgbotapi.types.ReplyInfo
 import kotlinx.serialization.Serializable
 
 @Serializable
-class Story : ExternalReplyInfo.ContentVariant
+class Story : ReplyInfo.External.ContentVariant
