+++ tgbotapi.core/src/commonMain/kotlin/dev/inmo/tgbotapi/utils/extensions/MediaGroupContentMessageCreator.kt
@@ -40,7 +40,7 @@ fun <T : MediaGroupPartContent> List<PossiblySentViaBotCommonMessage<T>>.asMedia
             sourceMessage.editDate,
             sourceMessage.hasProtectedContent,
             sourceMessage.forwardOrigin,
-            sourceMessage.replyTo,
+            sourceMessage.replyInfo,
             sourceMessage.replyMarkup,
             sourceMessage.senderBot,
             sourceMessage.authorSignature,
@@ -55,7 +55,7 @@ fun <T : MediaGroupPartContent> List<PossiblySentViaBotCommonMessage<T>>.asMedia
             sourceMessage.editDate,
             sourceMessage.hasProtectedContent,
             sourceMessage.forwardOrigin,
-            sourceMessage.replyTo,
+            sourceMessage.replyInfo,
             sourceMessage.replyMarkup,
             sourceMessage.senderBot,
             sourceMessage.mediaGroupId
@@ -67,7 +67,7 @@ fun <T : MediaGroupPartContent> List<PossiblySentViaBotCommonMessage<T>>.asMedia
             sourceMessage.forwardOrigin,
             sourceMessage.editDate,
             sourceMessage.hasProtectedContent,
-            sourceMessage.replyTo,
+            sourceMessage.replyInfo,
             sourceMessage.replyMarkup,
             content,
             sourceMessage.senderBot,
@@ -82,7 +82,7 @@ fun <T : MediaGroupPartContent> List<PossiblySentViaBotCommonMessage<T>>.asMedia
             sourceMessage.forwardOrigin,
             sourceMessage.editDate,
             sourceMessage.hasProtectedContent,
-            sourceMessage.replyTo,
+            sourceMessage.replyInfo,
             sourceMessage.replyMarkup,
             content,
             sourceMessage.senderBot,
@@ -96,7 +96,7 @@ fun <T : MediaGroupPartContent> List<PossiblySentViaBotCommonMessage<T>>.asMedia
             sourceMessage.forwardOrigin,
             sourceMessage.editDate,
             sourceMessage.hasProtectedContent,
-            sourceMessage.replyTo,
+            sourceMessage.replyInfo,
             sourceMessage.replyMarkup,
             content,
             sourceMessage.senderBot,
@@ -111,7 +111,7 @@ fun <T : MediaGroupPartContent> List<PossiblySentViaBotCommonMessage<T>>.asMedia
             sourceMessage.forwardOrigin,
             sourceMessage.editDate,
             sourceMessage.hasProtectedContent,
-            sourceMessage.replyTo,
+            sourceMessage.replyInfo,
             sourceMessage.replyMarkup,
             content,
             sourceMessage.senderBot,
@@ -126,7 +126,7 @@ fun <T : MediaGroupPartContent> List<PossiblySentViaBotCommonMessage<T>>.asMedia
             sourceMessage.forwardOrigin,
             sourceMessage.editDate,
             sourceMessage.hasProtectedContent,
-            sourceMessage.replyTo,
+            sourceMessage.replyInfo,
             sourceMessage.replyMarkup,
             content,
             sourceMessage.senderBot,
@@ -142,7 +142,7 @@ fun <T : MediaGroupPartContent> List<PossiblySentViaBotCommonMessage<T>>.asMedia
             sourceMessage.forwardOrigin,
             sourceMessage.editDate,
             sourceMessage.hasProtectedContent,
-            sourceMessage.replyTo,
+            sourceMessage.replyInfo,
             sourceMessage.replyMarkup,
             content,
             sourceMessage.senderBot,
@@ -157,7 +157,7 @@ fun <T : MediaGroupPartContent> List<PossiblySentViaBotCommonMessage<T>>.asMedia
             sourceMessage.forwardOrigin,
             sourceMessage.editDate,
             sourceMessage.hasProtectedContent,
-            sourceMessage.replyTo,
+            sourceMessage.replyInfo,
             sourceMessage.replyMarkup,
             content,
             sourceMessage.senderBot,
